Changed some things, details in description:
*Changelog*
-Removed StandCapabilityStorage.java
-Added CapabilityInit.java to reduce clutter in the main class
