+++ src/main/java/com/novarch/jojomod/gui/StandGUI.java
@@ -1,6 +1,5 @@
 package com.novarch.jojomod.gui;
 
-import com.novarch.jojomod.capabilities.stand.IStand;
 import com.novarch.jojomod.capabilities.stand.JojoProvider;
 import com.novarch.jojomod.util.JojoLibs;
 import net.minecraft.client.Minecraft;
@@ -10,23 +9,13 @@ public class StandGUI extends AbstractGui
 {
    public static final Minecraft mc = Minecraft.getInstance();
 
-   public void render()
-   {
-       if (mc == null)
-           return;
-       IStand props = JojoProvider.get(mc.player);
-       String text = "initial";
-       text = Integer.toString(props.getTimeLeft());
-       drawString(mc.fontRenderer, text, 4, 4, 0xFFFFFF);
-   }
-
-    public void renderMadeInHeaven(int value)
+    public void renderTimeValue(int ticks)
     {
-        if(value > 0)
+        if(ticks > 0)
         {
             String text = "";
-            int minutes = value / 1200;
-            int seconds = (value / 20) - (minutes * 60);
+            int minutes = ticks / 1200;
+            int seconds = (ticks / 20) - (minutes * 60);
             text = String.valueOf(minutes);
             if (seconds < 10)
                 text += ":0" + seconds;
@@ -36,24 +25,12 @@ public class StandGUI extends AbstractGui
         } else { drawString(mc.fontRenderer, "\"Heaven\" has begun!", 4, 4, 0xFFFFFF); }
     }
 
-    public void renderMadeInHeaven()
+    public void render()
     {
         JojoProvider.getLazy(Minecraft.getInstance().player).ifPresent(props ->  {
             int value = props.getTimeLeft();
             if (props.getStandOn() && props.getStandID() == JojoLibs.StandID.madeInHeaven) {
-                if (value > 0) {
-                    String text = "";
-                    int minutes = value / 1200;
-                    int seconds = (value / 20) - (minutes * 60);
-                    text = String.valueOf(minutes);
-                    if (seconds < 10)
-                        text += ":0" + seconds;
-                    else
-                        text += ":" + seconds;
-                    drawString(mc.fontRenderer, text, 4, 4, 0xFFFFFF);
-                } else {
-                    drawString(mc.fontRenderer, "\"Heaven\" has begun!", 4, 4, 0xFFFFFF);
-                }
+                renderTimeValue(value);
             }
         });
     }
